bridge/requirements.txt
cd /home/luca/petsafe_st/petsafe_smartfeed
echo "flask
pyyaml
requests
petsafe-smartfeed" > bridge/requirements.txt